requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
django-filter
djangorestframework_simplejwt
requests
